requirements.txt
streamlit
yfinance
requests
feedparser
PyPDF2
transformers
sentence-transformers
faiss-cpu
plotly
pandas
os
datetime
fitz
sumy
nltk
time
textblob
torch
numpy
cohere
matplotlib.pylot